+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.3.0
 -----
 
+ * Using the `format` option of `DateType` and `DateTimeType` when the `html5` option is enabled is deprecated.
  * Using names for buttons that do not start with a letter, a digit, or an underscore is deprecated and will lead to an
    exception in 5.0.
  * Using names for buttons that do not contain only letters, digits, underscores, hyphens, and colons is deprecated and